requirements.txt
alembic>=1.13.2,<2
fastapi>=0.115.0,<1
flask>=3.0.3,<4
httpx>=0.27.2,<1
joblib>=1.4.2,<2
numpy>=2.0.2,<3
pandas>=2.2.2,<3
psycopg[binary]>=3.2.1,<4
python-dotenv>=1.0.1,<2
scikit-learn>=1.5.1,<2
sqlalchemy>=2.0.32,<3
tenacity>=9.0.0,<10
uvicorn[standard]>=0.30.6,<1
markdown>=3.6,<4
bleach>=6.1.0,<7
